+++ src/shared_lib.rs
@@ -16,8 +16,9 @@ use crate::os::windows::shared_lib::{
 };
 use crate::{
     core::{
-        cstr::{nstd_core_cstr_as_ptr, nstd_core_cstr_get_null, NSTDCStr},
+        cstr::{nstd_core_cstr_as_ptr, nstd_core_cstr_get_null},
         optional::NSTDOptional,
+        str::{nstd_core_str_as_cstr, NSTDStr},
     },
     cstring::{nstd_cstring_as_ptr, nstd_cstring_from_cstr_unchecked},
     NSTDAny, NSTDAnyMut, NSTDChar,
@@ -39,7 +40,7 @@ pub type NSTDOptionalSharedLib = NSTDOptional<NSTDSharedLib>;
 ///
 /// # Parameters:
 ///
-/// - `const NSTDCStr *path` - A path to the shared library.
+/// - `const NSTDStr *path` - A path to the shared library.
 ///
 /// # Returns
 ///
@@ -55,11 +56,12 @@ pub type NSTDOptionalSharedLib = NSTDOptional<NSTDSharedLib>;
 ///
 /// - The loaded library may have platform-specific initialization routines ran when it is loaded.
 #[cfg_attr(feature = "clib", no_mangle)]
-pub unsafe extern "C" fn nstd_shared_lib_load(path: &NSTDCStr) -> NSTDOptionalSharedLib {
+pub unsafe extern "C" fn nstd_shared_lib_load(path: &NSTDStr) -> NSTDOptionalSharedLib {
     // Check if `path` is already null terminated.
-    if nstd_core_cstr_get_null(path).is_null() {
+    let path = nstd_core_str_as_cstr(path);
+    if nstd_core_cstr_get_null(&path).is_null() {
         // Allocate a null byte for `path`.
-        let path = nstd_cstring_from_cstr_unchecked(path);
+        let path = nstd_cstring_from_cstr_unchecked(&path);
         #[cfg(unix)]
         return nstd_os_unix_shared_lib_load(nstd_cstring_as_ptr(&path));
         #[cfg(windows)]
@@ -67,9 +69,9 @@ pub unsafe extern "C" fn nstd_shared_lib_load(path: &NSTDCStr) -> NSTDOptionalSh
     } else {
         // Use the already null terminated `path`.
         #[cfg(unix)]
-        return nstd_os_unix_shared_lib_load(nstd_core_cstr_as_ptr(path));
+        return nstd_os_unix_shared_lib_load(nstd_core_cstr_as_ptr(&path));
         #[cfg(windows)]
-        return nstd_os_windows_shared_lib_load(nstd_core_cstr_as_ptr(path));
+        return nstd_os_windows_shared_lib_load(nstd_core_cstr_as_ptr(&path));
     }
 }